Add another log with another full term extraction
This new log file is the output of a bug fix in the script where it was
finding one word segments in the file. I kept the file because it is
still good data. This new log file now checks that every token in the
list has alpha chars and then adds to corpus. Much bigger data set and
does aligned term extract on the segments.